requirements.txt
flask
flask-cors
python-dotenv
langgraph
rich
twilio
openai